refactor container pull function
