+++ pkg/container/auth.go
@@ -1,0 +1,21 @@
+package container
+
+import (
+	"encoding/base64"
+	"encoding/json"
+
+	"github.com/docker/docker/api/types/registry"
+)
+
+func GetRegistryAuth(username, password string) string {
+	authConfig := registry.AuthConfig{
+		Username: username,
+		Password: password,
+	}
+	encodedJSON, err := json.Marshal(authConfig)
+	if err != nil {
+		panic(err)
+	}
+	authStr := base64.URLEncoding.EncodeToString(encodedJSON)
+	return authStr
+}
